requirements.txt
aiogram>=3.2.0,<4.0.0
gspread>=5.0.0,<7.0.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.7.0
google-genai>=1.0.0 